parser/requirements.txt
aiohttp
playwright
requests
pillow
